+++ NEWS.md
@@ -1,5 +1,6 @@
 # resios (development version)
 
+* Fixed examples when run without token
 * Moved to rOpenSpain (documentation)
 * Initial release.